[WIP]: Refactoring `createItem` API for CRUD operations
This branch tracks the work in progress for refactoring how we use
`keystone.createItems` API. The existing API, allows the end user to seed
the data while setting up the keystone app. Doing so, the API bypass the
GraphQL layer and insert the data directly into the database.

This behaviour deviates from how we are using the executeGraphQl method, which
relies on the existing API we get from `schema => ({api, AdminUI})`.

So, this branch is an effort to provide an API that follows the
single API standard which is less ambiguous and provide a good developer
experience.